:recycle: api: move get-next-challenge-for-certification to src
