+++ api/tests/certification/course/unit/domain/usecases/get-next-challenge-for-certification_test.js
@@ -1,9 +1,9 @@
-import { domainBuilder, expect, sinon, catchErr } from '../../../test-helper.js';
-import { getNextChallengeForCertification } from '../../../../lib/domain/usecases/get-next-challenge-for-certification.js';
-import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
-import { CertificationVersion } from '../../../../src/shared/domain/models/CertificationVersion.js';
-import { AssessmentEndedError } from '../../../../lib/domain/errors.js';
-import { config } from '../../../../src/shared/config.js';
+import { domainBuilder, expect, sinon, catchErr } from '../../../../../test-helper.js';
+import { getNextChallengeForCertification } from '../../../../../../src/certification/course/domain/usecases/get-next-challenge-for-certification.js';
+import { Assessment } from '../../../../../../src/shared/domain/models/Assessment.js';
+import { CertificationVersion } from '../../../../../../src/shared/domain/models/CertificationVersion.js';
+import { AssessmentEndedError } from '../../../../../../lib/domain/errors.js';
+import { config } from '../../../../../../src/shared/config.js';
 
 describe('Unit | Domain | Use Cases | get-next-challenge-for-certification', function () {
   describe('#getNextChallengeForCertification', function () {
